+++ StardewXnbHack/Framework/Writers/BaseAssetWriter.cs
@@ -1,3 +1,6 @@
+using System;
+using Force.DeepCloner;
+using Newtonsoft.Json;
 using StardewModdingAPI.Toolkit.Serialization;
 using StardewModdingAPI.Toolkit.Utilities;
 
@@ -9,8 +12,8 @@ namespace StardewXnbHack.Framework.Writers
         /*********
         ** Private methods
         *********/
-        /// <summary>The underlying JSON helper preconfigured with serialization rules for Stardew Valley.</summary>
-        private readonly JsonHelper JsonHelper = new();
+        /// <summary>The settings to use when serializing JSON.</summary>
+        private static readonly Lazy<JsonSerializerSettings> JsonSettings = new(BaseAssetWriter.GetJsonSerializerSettings);
 
 
         /*********
@@ -37,13 +40,24 @@ namespace StardewXnbHack.Framework.Writers
         /// <param name="asset">The asset to serialize.</param>
         protected string FormatData(object asset)
         {
-            return this.JsonHelper.Serialize(asset);
+            return JsonConvert.SerializeObject(asset, BaseAssetWriter.JsonSettings.Value);
         }
 
         /// <summary>Get the recommended file extension for a data file formatted with <see cref="FormatData"/>.</summary>
         protected string GetDataExtension()
         {
             return "json";
+        }
+
+        /// <summary>Get the serializer settings to apply when writing JSON.</summary>
+        private static JsonSerializerSettings GetJsonSerializerSettings()
+        {
+            JsonHelper jsonHelper = new();
+            JsonSerializerSettings settings = jsonHelper.JsonSettings.DeepClone();
+
+            settings.ContractResolver = new IgnoreDefaultOptionalPropertiesResolver();
+
+            return settings;
         }
     }
 }